Extract engine text overlay operation
add_text is a focused operation family with drawtext validation, escaping, and output probing. This moves it into engine_text.py while re-exporting add_text from mcp_video.engine for existing callers.

Constraint: Preserve drawtext escaping, lazy facade compatibility, and EditResult output behavior.

Rejected: Move add_audio with add_text | audio handling uses different probing/mix behavior and should be reviewed separately.

Confidence: high

Scope-risk: narrow

Directive: Keep engine.py as compatibility facade while moving one operation family at a time.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_text.py

Tested: /opt/homebrew/bin/python3 add_text re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'add_text or text' tests/test_server.py -k 'add_text' tests/test_e2e.py -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
